+++ src/arch/mod.rs
@@ -1,0 +1,5 @@
+mod aarch64;
+mod arm;
+mod powerpc;
+mod wasm;
+mod x86_64;